+++ src/mbgl/renderer/symbol_bucket.hpp
@@ -1,22 +1,15 @@
 #pragma once
 
 #include <mbgl/renderer/bucket.hpp>
-#include <mbgl/tile/geometry_tile_data.hpp>
 #include <mbgl/map/mode.hpp>
-#include <mbgl/geometry/vao.hpp>
 #include <mbgl/geometry/elements_buffer.hpp>
 #include <mbgl/geometry/text_buffer.hpp>
 #include <mbgl/geometry/icon_buffer.hpp>
 #include <mbgl/geometry/collision_box_buffer.hpp>
-#include <mbgl/text/glyph.hpp>
-#include <mbgl/text/collision_feature.hpp>
-#include <mbgl/text/shaping.hpp>
-#include <mbgl/text/quads.hpp>
-#include <mbgl/style/filter.hpp>
+#include <mbgl/text/glyph_range.hpp>
 #include <mbgl/style/layers/symbol_layer_properties.hpp>
 
 #include <memory>
-#include <map>
 #include <set>
 #include <vector>
 
@@ -25,49 +18,13 @@ namespace mbgl {
 class SDFShader;
 class IconShader;
 class CollisionBoxShader;
-class CollisionTile;
-class SpriteAtlas;
-class SpriteStore;
-class GlyphAtlas;
-class GlyphStore;
-class IndexedSubfeature;
-
-class SymbolFeature {
-public:
-    GeometryCollection geometry;
-    std::u32string label;
-    std::string sprite;
-    std::size_t index;
-};
-
-struct Anchor;
-
-class SymbolInstance {
-    public:
-        explicit SymbolInstance(Anchor& anchor, const GeometryCoordinates& line,
-                const Shaping& shapedText, const PositionedIcon& shapedIcon,
-                const style::SymbolLayoutProperties&, const bool inside, const uint32_t index,
-                const float textBoxScale, const float textPadding, style::SymbolPlacementType textPlacement,
-                const float iconBoxScale, const float iconPadding, style::SymbolPlacementType iconPlacement,
-                const GlyphPositions& face, const IndexedSubfeature& indexedfeature);
-        Point<float> point;
-        uint32_t index;
-        bool hasText;
-        bool hasIcon;
-        SymbolQuads glyphQuads;
-        SymbolQuads iconQuads;
-        CollisionFeature textCollisionFeature;
-        CollisionFeature iconCollisionFeature;
-};
 
 class SymbolBucket : public Bucket {
-    typedef ElementGroup<2> TextElementGroup;
-    typedef ElementGroup<4> IconElementGroup;
-    typedef ElementGroup<1> CollisionBoxElementGroup;
-
 public:
-    SymbolBucket(uint32_t overscaling, float zoom, const MapMode, std::string bucketName_, std::string sourceLayerName_);
-    ~SymbolBucket() override;
+    SymbolBucket(const MapMode,
+                 style::SymbolLayoutProperties,
+                 bool sdfIcons,
+                 bool iconsNeedLinear);
 
     void upload(gl::ObjectStore&, gl::Config&) override;
     void render(Painter&, PaintParameters&, const style::Layer&, const RenderTile&) override;
@@ -77,81 +34,39 @@ public:
     bool hasCollisionBoxData() const;
     bool needsClipping() const override;
 
-    void addFeatures(uintptr_t tileUID,
-                     SpriteAtlas&,
-                     GlyphAtlas&,
-                     GlyphStore&);
-
     void drawGlyphs(SDFShader&, gl::ObjectStore&, PaintMode);
     void drawIcons(SDFShader&, gl::ObjectStore&, PaintMode);
     void drawIcons(IconShader&, gl::ObjectStore&, PaintMode);
     void drawCollisionBoxes(CollisionBoxShader&, gl::ObjectStore&);
 
-    void parseFeatures(const GeometryTileLayer&, const style::Filter&);
-    bool needsDependencies(GlyphStore&, SpriteStore&);
-    void placeFeatures(CollisionTile&) override;
-
-private:
-    void addFeature(const GeometryCollection &lines,
-            const Shaping &shapedText, const PositionedIcon &shapedIcon,
-            const GlyphPositions &face,
-            const size_t index);
-    bool anchorIsTooClose(const std::u32string &text, const float repeatDistance, Anchor &anchor);
-    std::map<std::u32string, std::vector<Anchor>> compareText;
-
-    void addToDebugBuffers(CollisionTile &collisionTile);
-
-    void swapRenderData() override;
-
-    // Adds placed items to the buffer.
-    template <typename Buffer, typename GroupType>
-    void addSymbols(Buffer &buffer, const SymbolQuads &symbols, float scale,
-            const bool keepUpright, const style::SymbolPlacementType placement, const float placementAngle);
-
-public:
-    style::SymbolLayoutProperties layout;
-
-    float iconMaxSize = 1.0f;
-    float textMaxSize = 16.0f;
-
-    bool sdfIcons = false;
-    bool iconsNeedLinear = false;
-
-private:
-
-    const float overscaling;
-    const float zoom;
-    const uint32_t tileSize;
-    const float tilePixelRatio;
     const MapMode mode;
-    const std::string bucketName;
-    const std::string sourceLayerName;
+    const style::SymbolLayoutProperties layout;
+    const bool sdfIcons;
+    const bool iconsNeedLinear;
 
-    std::set<GlyphRange> ranges;
-    std::vector<SymbolInstance> symbolInstances;
-    std::vector<SymbolFeature> features;
+private:
+    friend class SymbolLayout;
 
-    struct SymbolRenderData {
-        struct TextBuffer {
-            TextVertexBuffer vertices;
-            TriangleElementsBuffer triangles;
-            std::vector<std::unique_ptr<TextElementGroup>> groups;
-        } text;
+    typedef ElementGroup<2> TextElementGroup;
+    typedef ElementGroup<4> IconElementGroup;
+    typedef ElementGroup<1> CollisionBoxElementGroup;
 
-        struct IconBuffer {
-            IconVertexBuffer vertices;
-            TriangleElementsBuffer triangles;
-            std::vector<std::unique_ptr<IconElementGroup>> groups;
-        } icon;
+    struct TextBuffer {
+        TextVertexBuffer vertices;
+        TriangleElementsBuffer triangles;
+        std::vector<std::unique_ptr<TextElementGroup>> groups;
+    } text;
 
-        struct CollisionBoxBuffer {
-            CollisionBoxVertexBuffer vertices;
-            std::vector<std::unique_ptr<CollisionBoxElementGroup>> groups;
-        } collisionBox;
-    };
+    struct IconBuffer {
+        IconVertexBuffer vertices;
+        TriangleElementsBuffer triangles;
+        std::vector<std::unique_ptr<IconElementGroup>> groups;
+    } icon;
 
-    std::unique_ptr<SymbolRenderData> renderData;
-    std::unique_ptr<SymbolRenderData> renderDataInProgress;
+    struct CollisionBoxBuffer {
+        CollisionBoxVertexBuffer vertices;
+        std::vector<std::unique_ptr<CollisionBoxElementGroup>> groups;
+    } collisionBox;
 };
 
 } // namespace mbgl
