+++ src/mbgl/tile/tile_worker.hpp
@@ -4,7 +4,6 @@
 #include <mbgl/tile/tile_id.hpp>
 #include <mbgl/util/noncopyable.hpp>
 #include <mbgl/util/variant.hpp>
-#include <mbgl/text/placement_config.hpp>
 #include <mbgl/geometry/feature_index.hpp>
 
 #include <atomic>
@@ -22,10 +21,11 @@ class SpriteStore;
 class GlyphAtlas;
 class GlyphStore;
 class Bucket;
+class SymbolLayout;
+class PlacementConfig;
 
 namespace style {
 class Layer;
-class SymbolLayer;
 } // namespace style
 
 // We're using this class to shuttle the resulting buckets from the worker thread to the MapContext
@@ -42,6 +42,12 @@ using TileParseResult = variant<
     TileParseResultData, // success
     std::exception_ptr>;    // error
 
+class TilePlacementResult {
+public:
+    std::unordered_map<std::string, std::unique_ptr<Bucket>> buckets;
+    std::unique_ptr<CollisionTile> collisionTile;
+};
+
 class TileWorker : public util::noncopyable {
 public:
     TileWorker(OverscaledTileID,
@@ -54,21 +60,18 @@ public:
 
     TileParseResult parseAllLayers(std::vector<std::unique_ptr<style::Layer>>,
                                    std::unique_ptr<const GeometryTileData>,
-                                   PlacementConfig);
+                                   const PlacementConfig&);
 
-    TileParseResult parsePendingLayers(PlacementConfig);
+    TileParseResult parsePendingLayers(const PlacementConfig&);
 
     TileParseResult redoLayout(std::vector<std::unique_ptr<style::Layer>>,
-                               PlacementConfig);
+                               const PlacementConfig&);
 
-    std::unique_ptr<CollisionTile> redoPlacement(const std::unordered_map<std::string, std::unique_ptr<Bucket>>*,
-                       PlacementConfig);
+    TilePlacementResult redoPlacement(const PlacementConfig&);
 
 private:
-    TileParseResult prepareResult(const PlacementConfig& config);
-    void parseLayer(const style::Layer*);
-    void insertBucket(const std::string& name, std::unique_ptr<Bucket>);
-    std::unique_ptr<CollisionTile> placeLayers(PlacementConfig);
+    TileParseResult parsePendingLayers(const PlacementConfig&,
+                                       std::unordered_map<std::string, std::unique_ptr<Bucket>>);
 
     const OverscaledTileID id;
 
@@ -78,23 +81,11 @@ private:
     const std::atomic<bool>& obsolete;
     const MapMode mode;
 
-    bool partialParse = false;
-
     std::vector<std::unique_ptr<style::Layer>> layers;
-
-    std::unique_ptr<FeatureIndex> featureIndex;
     std::unique_ptr<const GeometryTileData> tileData;
 
-    // Contains buckets that we couldn't parse so far due to missing resources.
-    // They will be attempted on subsequent parses.
-    std::list<std::pair<const style::SymbolLayer*, std::unique_ptr<Bucket>>> pending;
-
-    // Contains buckets that have been parsed, but still need placement.
-    // They will be placed when all buckets have been parsed.
-    std::unordered_map<std::string, std::unique_ptr<Bucket>> placementPending;
-
-    // Temporary holder
-    TileParseResultData result;
+    std::unique_ptr<FeatureIndex> featureIndex;
+    std::vector<std::unique_ptr<SymbolLayout>> symbolLayouts;
 };
 
 } // namespace mbgl
