+++ src/mbgl/util/worker.cpp
@@ -63,10 +63,9 @@ public:
     }
 
     void redoPlacement(TileWorker* worker,
-                       const std::unordered_map<std::string, std::unique_ptr<Bucket>>* buckets,
                        PlacementConfig config,
-                       std::function<void(std::unique_ptr<CollisionTile>)> callback) {
-        callback(worker->redoPlacement(buckets, config));
+                       std::function<void(TilePlacementResult)> callback) {
+        callback(worker->redoPlacement(config));
     }
 };
 
@@ -120,12 +119,11 @@ Worker::redoLayout(TileWorker& worker,
 
 std::unique_ptr<AsyncRequest>
 Worker::redoPlacement(TileWorker& worker,
-                      const std::unordered_map<std::string, std::unique_ptr<Bucket>>& buckets,
                       PlacementConfig config,
-                      std::function<void(std::unique_ptr<CollisionTile>)> callback) {
+                      std::function<void(TilePlacementResult)> callback) {
     current = (current + 1) % threads.size();
     return threads[current]->invokeWithCallback(&Worker::Impl::redoPlacement, &worker,
-                                                &buckets, config, callback);
+                                                config, callback);
 }
 
 } // end namespace mbgl
